refactor(hush-cli): split pkg_cli install.rs registry path into submodule
install.rs was 614 lines — marginally over the ~600 decomposition target
and not on the plan's sanctioned-large list. Split along the existing clean
seam: relocate the registry-install path (select_default_registry_version,
read_response_bytes_limited, cmd_pkg_install_registry) into a new
pkg_cli/install/registry.rs child module, leaving the local-install,
identity, and rollback helpers in install/mod.rs.

Pure relocation: function bodies are byte-for-byte identical. The only
non-move changes are the module scaffolding (mod decl + re-export), use-block
narrowing for imports that followed the moved code, and one sanctioned
fn -> pub(super) fn widening on cmd_pkg_install_registry so the parent
router can call it across the module boundary. select_default_registry_version
keeps its exact pkg_cli-wide reach via pub(in crate::pkg_cli) and a
#[cfg(test)] re-export, matching its test consumer in pkg_cli/mod.rs.

Resulting files: install/mod.rs 258 lines, install/registry.rs 378 lines —
both comfortably under 600. No public API or test changed.
